✉ condense cmds.ts -> bot.ts
