+++ discord/package.json
@@ -5,7 +5,6 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "cmds": "tsx watch --env-file=.env src/commands.ts",
     "bot": "tsx watch --env-file=.env src/bot.ts"
   },
   "keywords": [],
