--- discord/src/commands.ts
@@ -1,34 +1,0 @@
-import { REST, Routes } from 'discord.js';
-
-const { BOT_TOKEN, APPLICATION_ID } = process.env;
-
-if (!BOT_TOKEN || !APPLICATION_ID) {
-  throw new Error(
-    'BOT_TOKEN and APPLICATION_ID must be defined in a supplied .env file.'
-  );
-}
-
-const commands = [
-  {
-    name: 'ping',
-    description: 'Replies with Pong!',
-  },
-  {
-    name: 'newevent',
-    description: 'Creates a new event',
-  },
-];
-
-const rest = new REST({ version: '10' }).setToken(BOT_TOKEN);
-
-try {
-  console.log('Started refreshing application (/) commands.');
-
-  await rest.put(Routes.applicationCommands(APPLICATION_ID), {
-    body: commands,
-  });
-
-  console.log('Successfully reloaded application (/) commands.');
-} catch (error) {
-  console.error(error);
-}
